Log build version, as included in the template
