+++ core/src/main/java/io/questdb/log/LogAlertSocketWriter.java
@@ -24,6 +24,7 @@
 
 package io.questdb.log;
 
+import io.questdb.BuildInformationHolder;
 import io.questdb.mp.QueueConsumer;
 import io.questdb.mp.RingQueue;
 import io.questdb.mp.SCSequence;
@@ -46,6 +47,7 @@ public class LogAlertSocketWriter extends SynchronizedJob implements Closeable, 
     static final CharSequenceObjHashMap<CharSequence> ALERT_PROPS = TemplateParser.adaptMap(System.getenv());
     private static final String DEFAULT_ENV_VALUE = "GLOBAL";
     private static final String ORG_ID_ENV = "ORGID";
+    private static final String QDB_VERSION_ENV = "QDB_VERSION";
     private static final String NAMESPACE_ENV = "NAMESPACE";
     private static final String CLUSTER_ENV = "CLUSTER_NAME";
     private static final String INSTANCE_ENV = "INSTANCE_NAME";
@@ -345,6 +347,9 @@ public class LogAlertSocketWriter extends SynchronizedJob implements Closeable, 
         }
         if (!ALERT_PROPS.contains(INSTANCE_ENV)) {
             ALERT_PROPS.put(INSTANCE_ENV, DEFAULT_ENV_VALUE);
+        }
+        if (!ALERT_PROPS.contains(QDB_VERSION_ENV)) {
+            ALERT_PROPS.put(QDB_VERSION_ENV, BuildInformationHolder.INSTANCE.toString());
         }
         ALERT_PROPS.put(MESSAGE_ENV, MESSAGE_ENV_VALUE);
     }
