+++ core/src/main/resources/alert-manager-tpt.json
@@ -6,6 +6,7 @@
       "service": "QuestDB",
       "category": "application-logs",
       "severity": "critical",
+      "version": "${QDB_VERSION}",
       "cluster": "${CLUSTER_NAME}",
       "orgid": "${ORGID}",
       "namespace": "${NAMESPACE}",
